+++ src/ska_oso_services/odt/api/prsls.py
@@ -97,8 +97,15 @@ def prj_details() -> list[ProposalProjectDetails]:
             start=datetime.fromisoformat("1970-01-01T00:00:00.000000+00:00"),
         )
 
-        all_projects = uow.prjs.query(date_query)
+        def get_latest_entities(entity_list, id_field):
+            return list(
+                {
+                    getattr(entity, id_field): entity
+                    for entity in sorted(entity_list, key=lambda x: x.metadata.version)
+                }.values()
+            )
 
+        all_projects = get_latest_entities(uow.prjs.query(date_query), "prj_id")
         # As we only display the Proposal ID which is already present in the Project,
         # we can handle the Project with and without Proposal cases in the same
         # way here, without having to search through all_proposals for extra fields
@@ -116,7 +123,7 @@ def prj_details() -> list[ProposalProjectDetails]:
             for project in all_projects
         ]
 
-        all_proposals = uow.prsls.query(date_query)
+        all_proposals = get_latest_entities(uow.prsls.query(date_query), "prsl_id")
         # As the Project/Proposal relationship is stored as a foreign key
         # on the Project, to find Proposals without Projects we need to filter
         # in this roundabout way